feat(tags): enforce user ownership on tag relationships
feat(transactions): update transaction model and improve handling of amounts

refactor(toasts): centralize toast notifications for better management

fix(auth): streamline authentication error handling and session management

fix(transactions): ensure correct handling of transaction updates and deletions

+++ backend/database/migrations/2025_03_29_155905_create_transactions_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained()->cascadeOnDelete();
             $table->timestamp('at')->index();
-            $table->bigInteger('amount')->default(0);
+            $table->decimal('amount', 15, 2)->default(0);
             $table->string('note')->nullable();
             $table->timestamp('created_at')->useCurrent();
             $table->timestamp('updated_at')->useCurrent()->useCurrentOnUpdate();
